move helpers from `_numeric.py` to `_util.numeric_checks`
